backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlalchemy==2.0.41
pydantic==2.11.7
pydantic-core==2.33.2
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
email-validator==2.2.0
anyio==4.9.0
starlette==0.46.2
sniffio==1.3.1
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
pyasn1==0.6.1
rsa==4.9.1
six==1.17.0
greenlet==3.2.3
typing-extensions==4.14.0
annotated-types==0.7.0
typing-inspection==0.4.1
h11==0.16.0
MarkupSafe==3.0.2 
python-http-client==3.3.7 
sendgrid==6.12.4 
werkzeug==3.1.3
python-dotenv==1.1.0
